# src/models/openapi.schema.yaml
openapi: 3.0.3
info:
  version: '1.2'
  title: 'SCB Recipes API'
  description: 'Secret Cook Book Recipes REST API'
  license:
    name: MIT
    url: https://spdx.org/licenses/MIT
  contact:
    name: Cagil Seker
    url: https://github.com/cagils
    email: 'cagils@gmail.com'
paths:
  /my/recipes:
    get:
      summary: List recipes
      parameters: []
      responses:
        '200':
          description: successful
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Recipe'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Create recipe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Recipe'
      responses:
        '201':
          description: successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipe'
        '400':
          description: recipe data is invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /my/recipes/{recipeID}:
    get:
      summary: Get recipe
      parameters:
        - name: recipeID
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipe'
        '404':
          description: Recipe not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: Replace recipe
      parameters:
        - name: recipeID
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Recipe'
      responses:
        '200':
          description: successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipe'
        '400':
          description: Invalid recipe parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not authorized to update the recipe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      summary: Update recipe
      parameters:
        - name: recipeID
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecipeOptional'
      responses:
        '200':
          description: successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipe'
        '400':
          description: Invalid recipe parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not authorized to update the recipe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    Error:
      title: Error Schema
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
        message:
          type: string
    RecipeOptional:
      title: Recipe Schema (optional properties)
      type: object
      default: {}
      properties:
        recipeId:
          type: string
          default: ''
          example: scb0001
        userId:
          type: string
          default: ''
          example: user0001
        photo:
          type: string
          default: ''
          example: url
        title:
          type: string
          default: ''
          example: Easy chocolate fudge cake
        shortDesc:
          type: string
          default: ''
          example: Need a guaranteed crowd-pleasing cake that's easy to make? This super-squidgy chocolate fudge cake with smooth icing is an instant baking win
        serving:
          type: integer
          default: 0
          example: 8
        difficulty:
          type: integer
          default: 0
          example: 0
        time:
          type: integer
          default: 0
          example: 45
        ingredients:
          type: array
          default: []
          items:
            $ref: '#/components/schemas/IngredientGroup'
        description:
          type: object
          default: {}
          required:
            - html
            - json
          properties:
            html:
              type: string
              default: ''
              example: <h3><strong>Method</strong></h3><ul><li><p><strong>STEP 1</strong></p><p>Heat the oven to 180C/160C fan/gas 4. Oil and line the base of two 18cm sandwich tins. Sieve the flour, cocoa powder and bicarbonate of soda into a bowl. Add the caster sugar and mix well.</p></li><li><p><strong>STEP 2</strong></p><p>Make a well in the centre and add the golden syrup, eggs, sunflower oil and milk. Beat well with an electric whisk until smooth.</p></li><li><p><strong>STEP 3</strong></p><p>Pour the mixture into the two tins and bake for 25-30 mins until risen and firm to the touch. Remove from oven, leave to cool for 10 mins before turning out onto a cooling rack.</p></li><li><p><strong>STEP 4</strong></p><p>To make the icing, beat the unsalted butter in a bowl until soft. Gradually sieve and beat in the icing sugar and cocoa powder, then add enough of the milk to make the icing fluffy and spreadable.</p></li><li><p><strong>STEP 5</strong></p><p>Sandwich the two cakes together with the butter icing and cover the sides and the top of the cake with more icing.</p></li></ul>
            json:
              type: object
              default: {}
              properties: {}
    Recipe:
      allOf:
        - $ref: '#/components/schemas/RecipeOptional'
      title: Recipe Schema (required properties)
      required:
        - recipeId
        - userId
        - photo
        - title
        - shortDesc
        - serving
        - difficulty
        - time
        - ingredients
        - description
    IngredientGroup:
      title: Ingredient Group Schema
      type: object
      required:
        - groupName
        - list
      properties:
        groupName:
          type: string
          example: default
        list:
          type: array
          items:
            $ref: '#/components/schemas/Ingredient'
    Ingredient:
      title: The Ingredient Schema
      type: string
      example: 150ml sunflower oil, plus extra for the tin